Refactored out `getPoint` from components/content to utils so as to make it more reusable.
